Succesful Editor build on x64

+++ Mini_Maniacs/EditorSystem.cpp
@@ -126,7 +126,7 @@ void EditorSystem::Update(void)
     SaveData();
   }
 
-  if (InputSystem::isTriggered(Delete))
+  if (InputSystem::isTriggered(Delete) && Selected)
   {
     DeleteEntity();
   }
@@ -142,9 +142,9 @@ void EditorSystem::Update(void)
     {
     case entity:
       if(SelectedOBJ.OBJ.e->Parent() != nullptr)
-        SelectedOBJ.OBJ.e->GetComponent<Transform>()->SetLocalPosition(mousePos);
+        SelectedOBJ.OBJ.e->GetComponent<Transform>()->SetLocalPosition(mousePosGlobal);
       else
-        SelectedOBJ.OBJ.e->GetComponent<Transform>()->SetPosition(mousePos);
+        SelectedOBJ.OBJ.e->GetComponent<Transform>()->SetPosition(mousePosGlobal);
       scale = SelectedOBJ.OBJ.e->GetComponent<Transform>()->GetScale();
       break;
